datalib: class DataFrame added some FunctionColumn tests

+++ datalib/src/main/java/com/biorecorder/data/frame/Function.java
@@ -4,5 +4,5 @@ package com.biorecorder.data.frame;
  * Created by galafit on 11/3/19.
  */
 public interface Function {
-    int apply(double value);
+    double apply(double value);
 }
